requirements.txt
flask
mysql.connector
flask_mail
